[front] refacto: move transport selector to a new file
